Script to check cpuset.mems mismatch
Due to a race condition between Kubernetes and/or docker, and the
onlining of GPU memory, services may start without all GPUs causing
CUDA failures.

As described in https://bugzilla.redhat.com/show_bug.cgi?id=1746415
If Kubernetes, or others create their cgroup slice before the all GPU
Memory is available, they will have an incomplete list.  This causes
issues when a container has access to a GPU's device, but it's memory
does not show up in the cpuset.mems file.  Leading to a CUDA
intiliazation error

This script provides a workaround allowing users to effectively remove
the incorrect slices, and allow the services (docker, kubernetes, etc)
to rebuild with the correct values taken from the original cpuset.mems.