Add gpucheck to see if base cpuset.mems is ready
Add a check to the script to tell if the system is initially at a valid
memory state.  i.e. all gpu memory is onlined and added to cgroups.  The
check uses lspci to gather the number of V100 gpus in the box, and
calculate the correct cpuset.mems.  After the calculation a comparison
is made between the calculated cpuset.mems and the system created one.
If they match, we can safely assume all gpu nodes are onlined.  If not a
message is printed out and a Return Code of 1 is generated.