Add --use_move parameter when correcting slices.
In some cases the rmdir will not allow for the removal of the
slice directories.  If someone really wants the directory gone,
a mv will succeed in places where rmdir fails.  Most notably when
some process has the slice directory actively in use.

Now the --use_move can be used in conjuction with the --correct
parameter to allow for the removal command change.