Fix aif360 to include datasets
The aif360 includes 5 datasets as part of the pip package. As some of this data
is converted from one format to another after it is downloaded from the web, we
leverage the manual work done to create the pip package and just download the
pip package and convert it into a conda package.

Included the license agreements or citiations for each dataset

Copy the LICENSE file to the recipe directory as we no longer retrieve it from
the git source.