Remove h5py recipe
The h5py recipe version 2.9.0 for power only we added as a potential
workaround to a 1.5.4 issue. This workaround was never used and now
having h5py version 2.9.0 is causing a conflict with 1.7.0 if using
strict channel priority as 1.7.0 references h5py 2.8.0, which is not
available in this channel.

This will remove it from the conda-recipes, but it needs to be manually
removed from the powerai channel