add 'rust-nightly' feedstock based on conda-forge/rust-feedstock (#223)
Add a feedstock recipe for building a 'rust-nightly' package from the
upstream binary distribution tarballs. This recipe is based on:

https://github.com/conda-forge/rust-feedstock

And note some discussion there about building nightlies:

https://github.com/conda-forge/rust-feedstock/issues/31

One complication with the nightlies is that the source URLs for rust
nightly builds are only versioned with the word 'nightly'. They don't
contain an indication of the actual rust version that's included,
and presumably they'll change from time to time.

That leads to 2 issues:

1) The SHA hashes in the recipe may often be invalidated.

"Fix" this by removing the hashes from meta.yaml.

2) The 'rust-nightly' package version is hard-coded in meta.yaml and
so may fall out of sync with the internal rust version in the
download. In this case the build would succeed, but the package
version would be misleading.

Fix this by including a build test to confirm that the rust version
in the built package actually matches the package version. That
will at least force a build break, rather than allowing creation
of a mis-versioned package.