add allennlp feedstock
Add feedstock for allennlp, an open-source NLP research library,
built on PyTorch.

This feedstock is based on the existing conda-forge feedstock,
updated for a newer version (0.9.0 vs original 0.7.2) of allennlp.

See:

- https://github.com/conda-forge/allennlp-feedstock
- https://github.com/allenai/allennlp
- http://www.allennlp.org/

Include minor fixes for various unit tests (which are shipped
and can be run with: allennlp test-install).

Use the '1.7.0-pytorch' image since this package has a
pre-req ('pytorch-pretrained-bert') which in turn requires
PyTorch 1.3.1. And use the '-cpu' flavor our 1.7.0 images
since Travis doesn't have GPU (and so unit tests would fail
for lack of libcuda.so in some cases).