Fix ray import error
To fix the ray error with rpath, I moved the compilers from the build section of
requirements to the host section of requirements. The conda documentation says
that any shared libraries that you need to link against in your build should be
in the host section. This solved the rpath issue.

I reordered the build section to be the above the host section for to be
consistent across recipes.

This PR uses the common build_scripts now for ray-feedstock