Bazel creates the build artifacts without the user having write access to the file, when conda tries to modify the rpath it gets a permission denied error. To fix this problem we do a chmod -R u+w $PREFIX as the last steps of the build.sh script.
To solve the issue of python_abi being added as a dependency, had to remove conda-forge
from  the config files. Using python from conda-forge causes the python_abi dependency to
get added. Reordered the channel sources to the order we want.

Using the commaon build process, removed the build_scripts in ray-feedstock.