common build: add callout to feedstock pre-build script
Allow feedstocks to provide a script to be called just before
"conda build". This allows feedstocks to use the common build
scripts but further customize the build environment if needed.

The script must be named "build_steps_hook.sh" and must reside
in the feedstock's top-level directory (not the "recipe/"
directory, to avoid confusion with any scripts that "conda build"
will later run.)

The script is run with "bash ..." so should be a bash script,
but need not have file execute permissions..